Fixed TabNav view Reordered files locations
