+++ src/components/GlobalError.vue
