+++ src/components/Sensor.vue
