+++ src/components/sensors/Barometer.vue
