+++ src/components/sensors/BatteryLevel.vue
