+++ src/components/sensors/DefaultSensor.vue
