+++ src/components/sensors/LineChart.vue
