+++ src/components/sensors/Rain.vue
