+++ src/components/sensors/Thermometer.vue
