+++ src/components/sensors/chart-data.js
